+++ control/matlab/__init__.py
@@ -62,10 +62,10 @@ if not ('.config' in sys.modules):
 
 # Control system library
 from ..statesp import *
-from ..iosys import ss, rss, drss       # moved from .statesp
+from ..statesp import ss, rss, drss       # moved from .statesp
 from ..xferfcn import *
 from ..lti import *
-from ..namedio import *
+from ..iosys import *
 from ..frdata import *
 from ..dtime import *
 from ..exception import ControlArgument
